feature: CBT-4218: Review Markups
