+++ Atlas.ManualTesting/DependencyInjection/ServiceConfiguration.cs
@@ -47,8 +47,7 @@ namespace Atlas.ManualTesting.DependencyInjection
                 OptionsReaderFor<MatchingAlgorithmHttpFunctionSettings>(),
                 OptionsReaderFor<TopLevelHttpFunctionSettings>(),
                 OptionsReaderFor<PublicApiHttpFunctionSettings>(),
-                OptionsReaderFor<RepeatSearchHttpFunctionSettings>(),
-                OptionsReaderFor<SearchTrackingHttpFunctionSettings>());
+                OptionsReaderFor<RepeatSearchHttpFunctionSettings>());
         }
 
         private static void RegisterSettings(this IServiceCollection services)
@@ -65,7 +64,6 @@ namespace Atlas.ManualTesting.DependencyInjection
             services.RegisterAsOptions<RepeatSearchHttpFunctionSettings>("Debug:RepeatSearch");
             services.RegisterAsOptions<TopLevelHttpFunctionSettings>("Debug:TopLevel");
             services.RegisterAsOptions<PublicApiHttpFunctionSettings>("Debug:PublicApi");
-            services.RegisterAsOptions<SearchTrackingHttpFunctionSettings>("Debug:SearchTracking");
         }
 
         private static void RegisterServices(
