+++ Atlas.SearchTracking.Functions/Functions/SearchTrackingEventFunctions.cs
@@ -3,23 +3,17 @@ using Atlas.SearchTracking.Services;
 using Azure.Messaging.ServiceBus;
 using Microsoft.Azure.Functions.Worker;
 using System.Text;
-using Atlas.Debug.Client.Models.SearchTracking;
 using Atlas.SearchTracking.Common.Config;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Atlas.Client.Models.Search.Results;
 
 namespace Atlas.SearchTracking.Functions.Functions
 {
-    public class SearchTrackingFunctions
+    public class SearchTrackingEventFunctions
     {
         private readonly ISearchTrackingEventProcessor searchTrackingEventProcessor;
-        private readonly ISearchTrackingDebugService searchTrackingDebugService;
 
-        public SearchTrackingFunctions(ISearchTrackingEventProcessor searchTrackingEventProcessor, ISearchTrackingDebugService searchTrackingDebugService)
+        public SearchTrackingEventFunctions(ISearchTrackingEventProcessor searchTrackingEventProcessor)
         {
             this.searchTrackingEventProcessor = searchTrackingEventProcessor;
-            this.searchTrackingDebugService = searchTrackingDebugService;
         }
 
         [Function(nameof(HandleSearchTrackingEvent))]
@@ -34,17 +28,6 @@ namespace Atlas.SearchTracking.Functions.Functions
                 Enum.Parse<SearchTrackingEventType>(message.ApplicationProperties.GetValueOrDefault(SearchTrackingConstants.EventType).ToString());
 
             await searchTrackingEventProcessor.HandleEvent(body, eventType);
-        }
-
-        [Function(nameof(GetSearchRequestByIdentifier))]
-        public async Task<IActionResult> GetSearchRequestByIdentifier(
-            [HttpTrigger(AuthorizationLevel.Function, "get")]
-            HttpRequest request)
-        {
-            var searchIdentifier = Guid.Parse(request.Query["searchIdentifier"]);
-            var searchRequest = await searchTrackingDebugService.GetSearchRequestByIdentifier(searchIdentifier);
-
-            return new JsonResult(searchRequest);
         }
     }
 }
