+++ Atlas.SearchTracking.Functions/Program.cs
@@ -10,8 +10,7 @@ using Atlas.SearchTracking.Data.Repositories;
 using Atlas.SearchTracking.Services;
 using Microsoft.EntityFrameworkCore;
 using Atlas.SearchTracking.Common.Settings.ServiceBus;
-using AutoMapper;
-using Atlas.SearchTracking.Mapping;
+using Atlas.SearchTracking.Functions.Config;
 
 var host = new HostBuilder()
     .ConfigureFunctionsWebApplication(builder =>
@@ -42,14 +41,7 @@ var host = new HostBuilder()
         services.AddScoped<IMatchPredictionRepository, MatchPredictionRepository>();
         services.AddScoped<ISearchRequestMatchingAlgorithmAttemptsRepository, SearchRequestMatchingAlgorithmAttemptsRepository>();
         services.RegisterAsOptions<SearchTrackingServiceBusSettings>("MessagingServiceBus");
-
-        var mapperConfig = new MapperConfiguration(mc =>
-        {
-            mc.AddProfile(new SearchTrackingSearchRequestProfile());
-        });
-
-        IMapper mapper = mapperConfig.CreateMapper();
-        services.AddSingleton(mapper);
+        services.AddSingleton(sp => AutoMapperConfig.CreateMapper());
     })
     .Build();
